+ Finishing repository logic and starting tests

+++ app/src/main/java/com/dvidal/samplecurrencies/features/currencies/presentation/RateView.kt
@@ -13,7 +13,7 @@ data class RatesPresentationResponse(
 data class RatePresentation(
     val symbol: String,
     val rateName: Int? = null,
-    val value: Float? = null,
+    val value: Double? = null,
     val image: Int? = null,
     val isDefault: Boolean = false
 )
